tests/requirements.txt
pyserial>=3.5
pytest>=7.0
pytest-ordering>=0.6
